Save2.7

+++ main/src/main/java/ru/practicum/event/dto/EventShortDto.java
@@ -17,5 +17,5 @@ public class EventShortDto {
     private UserShortDto initiator;
     private Boolean paid;               //оплачиваемый?
     private String title;
-    private Long views;
+    private Integer views;
 }
